requirements.txt
aiofiles==0.6.0
appdirs==1.4.4
async-timeout==4.0.3
black==20.8b1
click==7.1.2
h11==0.12.0
hiredis==3.0.0
mypy-extensions==0.4.3
pathspec==0.8.1
redis==5.0.8
regex==2021.4.4
starlette==0.14.2
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
uvicorn==0.13.4
websockets==8.1
python-dotenv==1.0.1
requests
asgiref
# pydantic==2.8.2
